requirements.txt
Flask>=2.0
Werkzeug>=2.0
psycopg2-binary
gunicorn
python-dotenv
